serverutils: add StartTenant
Allow starting a tenant process pointing at a node from a TestCluster.
This paves the way to writing simple unit tests for multi-tenancy
from anywhere in the codebase.

The methods that construct and start the server are almost pure code
movement, with minimal adjustments.

Release note: None
